requirements.txt
asgiref==3.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-auth==1.5.1
dj-rest-auth==2.1.10
Django==3.2.8
django-allauth==0.44.0
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.3
Markdown==3.3.4
oauthlib==3.1.1
Pillow==9.0.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.1
urllib3==1.25.11
